refactor: split near-cap AI planning hosts and ratchet the 250-line lib gate (#4602)
Give Slice A of the phase-16 maintainability pass durable headroom by topic-splitting suppression, orchestrator input, lock-recovery scan, and essay-heavy filters, then enforce an empty 250-line AI lib allowlist.
