+++ SPEC/program/ai-source-file-size.md
@@ -1,7 +1,7 @@
 # AI package — lib/src physical line size ratchet (repo lint)
 
 **SPEC/program** — repository lint gate that keeps every hand-written Dart
-file under `packages/colonizethis_ai/lib/src/` at or below **300 physical
+file under `packages/colonizethis_ai/lib/src/` at or below **250 physical
 lines**, with an optional shrink-only grandfather allowlist.
 
 ## Motivation
@@ -13,11 +13,12 @@ concern-split the eight remaining near-cap planning modules and ratcheted
 the ceiling to **400**. Phase 13 Slice B (#4310) concern-split fourteen
 near-cap modules and ratcheted the ceiling to **350**. Phase 14 Slice A
 (#4365) concern-split fourteen near-cap planning modules (measured largest
-**316** physical) and ratcheted the ceiling to **300** so additive planning
-work retains durable headroom. This gate mirrors the save/data shrink-only
-allowlist pattern. AI is not an extracted logic-domain package, so it is
-not listed under `repo.domain_package_source_file_size` (still 500
-elsewhere).
+**316** physical) and ratcheted the ceiling to **300**. Phase 16 Slice A
+(#4602) concern-split remaining near-cap planning hosts and ratcheted the
+ceiling to **250** so additive planning work retains durable headroom. This
+gate mirrors the save/data shrink-only allowlist pattern. AI is not an
+extracted logic-domain package, so it is not listed under
+`repo.domain_package_source_file_size` (still 500 elsewhere).
 
 ## Source of truth
 
@@ -43,14 +44,14 @@ temporarily exceed the ceiling. The allowlist is **shrink-only**:
 - Path exists and is still over ceiling → skip size violation for that path.
 
 After #4079 Slice C / #4104 Slice B / #4239 Slice A / #4310 Slice B /
-#4365 Slice A the production allowlist is empty.
+#4365 Slice A / #4602 Slice A the production allowlist is empty.
 
 ## Acceptance criteria
 
 - Given every hand-written file under `packages/colonizethis_ai/lib/src/**` is
-  at or below 300 physical lines and the grandfather allowlist is empty, when
+  at or below 250 physical lines and the grandfather allowlist is empty, when
   the System runs `runCheckAiSourceFileSize`, then the System exits `0`.
-- Given a hand-written AI `lib/src` file with more than 300 physical lines and
+- Given a hand-written AI `lib/src` file with more than 250 physical lines and
   an empty grandfather allowlist, when the System runs `runCheckAiSourceFileSize`,
   then the System exits non-zero and names that file.
 - Given an over-cap AI `lib/src` file listed on the grandfather allowlist, when
@@ -62,6 +63,6 @@ After #4079 Slice C / #4104 Slice B / #4239 Slice A / #4310 Slice B /
 - Given a grandfather allowlist entry that names a file now at or below the
   ceiling, when the System runs `runCheckAiSourceFileSize`, then the System
   exits non-zero and reports that the entry must be removed from the allowlist.
-- Given a generated AI `lib/src` file (for example `*.g.dart`) over 300
+- Given a generated AI `lib/src` file (for example `*.g.dart`) over 250
   physical lines, when the System runs `runCheckAiSourceFileSize`, then the
   System does not fail because of that generated file.
