+++ SPEC/program/repo-lint.md
@@ -207,7 +207,7 @@
 | `tool/check_ai_contracts_h8_test_shared_fixtures.dart` | Forbid local `_flaggedSellerGame` redeclarations in the three H8 seller feedstock acquisition-target pins and local `_supplierGame` redeclarations in the OW feedstock prospect localization / mineral-eligibility pins under `packages/colonizethis_ai_contracts/test/`. Pins MUST import `flaggedSellerGame` / `supplierGame` from `test/support/h8_flagged_seller_game.dart` and `test/support/h8_supplier_prospect_game.dart`. The gate also fails when either shared support file is missing (Refs #4084); rule `repo.ai_contracts_h8_test_shared_fixtures` |
 | `tool/check_ai_contracts_source_file_size.dart` | Ratchet ceiling for `packages/colonizethis_ai_contracts/lib/**` physical lines per file (wave-2 ≤**260**; empty shrink-only grandfather; generated suffixes excluded); stricter than the shared 500-line `repo.domain_package_source_file_size` cap; rule `repo.ai_contracts_source_file_size` (Refs #4368) |
 | `tool/check_ai_contracts_test_file_size.dart` | Ratchet ceiling for `packages/colonizethis_ai_contracts/test/**` physical lines per file (wave-2 ≤**280**; empty shrink-only grandfather); stricter than `repo.domain_package_test_file_size` (400); rule `repo.ai_contracts_test_file_size` (Refs #4368) |
-| `tool/check_ai_source_file_size.dart` | Enforce `packages/colonizethis_ai/lib/src/**` Dart files at **300 physical lines or fewer**, with an optional shrink-only grandfather allowlist that fails when entries are missing or now under-cap (mirrors save/data size gates; headroom ratchet after #4104 Slice B, #4239 Slice A, #4310 Slice B, and #4365 Slice A near-cap splits; Refs #4079 Slice C, #4104 Slice B, #4239 Slice A, #4310, #4365) — see `SPEC/program/ai-source-file-size.md`; rule `repo.ai_source_file_size` |
+| `tool/check_ai_source_file_size.dart` | Enforce `packages/colonizethis_ai/lib/src/**` Dart files at **250 physical lines or fewer**, with an optional shrink-only grandfather allowlist that fails when entries are missing or now under-cap (mirrors save/data size gates; headroom ratchet after #4104 Slice B, #4239 Slice A, #4310 Slice B, #4365 Slice A, and #4602 Slice A near-cap splits; Refs #4079 Slice C, #4104 Slice B, #4239 Slice A, #4310, #4365, #4602) — see `SPEC/program/ai-source-file-size.md`; rule `repo.ai_source_file_size` |
 | `tool/check_models_test_file_size.dart` | Fail when any `.dart` file under `packages/colonizethis_models/test/**` exceeds 250 physical lines (wave-4 ratchet; Refs #4068 Slice D, #4571); rule `repo.models_test_file_size` |
 | `tool/check_models_lib_physical_file_size.dart` | Fail when any non-generated `.dart` file under `packages/colonizethis_models/lib/**` exceeds 250 physical lines (wave-4 ratchet 300→250; empty shrink-only grandfather; Refs #4334, #4571); rule `repo.models_lib_physical_file_size` |
 | `tool/check_models_test_mirrors_lib.dart` | Keep the `packages/colonizethis_models/test/**` tree mirrored to `lib/` per the testing rule (`test/` mirrors `lib/`): any Dart source that sits **directly** at the `test/` root is a violation — tests must live in subdirectories such as `test/src/`, `test/app_events/`, or `test/diplomacy/` that mirror `lib/src/` (Refs #4334); rule `repo.models_test_mirrors_lib` |
@@ -648,9 +648,9 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given every `.dart` file under `packages/colonizethis_ai_contracts/test/**` is at or below 280 physical lines and the grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_test_file_size`, then the rule passes without violations and exits `0` (Refs #4368).
 - Given a `packages/colonizethis_ai_contracts/test/**` Dart file with 281 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_test_file_size`, then the run fails naming that file and exits `1` (Refs #4368).
 - Given the `packages/colonizethis_ai_contracts/test` scan root is missing, when `repo.ai_contracts_test_file_size` runs, then the rule exits `1` (Refs #4368).
-- Given every hand-written file under `packages/colonizethis_ai/lib/src/**` is at or below 300 physical lines and the grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the rule passes without violations and exits `0` (Refs #4079, #4104, #4239, #4310, #4365).
-- Given a hand-written `packages/colonizethis_ai/lib/src/**` Dart file with more than 300 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the run fails naming that file and exits `1` (Refs #4079, #4104, #4239, #4310, #4365).
-- Given a grandfather allowlist entry for `repo.ai_source_file_size` that names a missing file or a file now at or below 300 physical lines, when the rule runs, then the run fails reporting a stale grandfather entry and exits `1` (Refs #4079, #4104, #4239, #4310, #4365).
+- Given every hand-written file under `packages/colonizethis_ai/lib/src/**` is at or below 250 physical lines and the grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the rule passes without violations and exits `0` (Refs #4079, #4104, #4239, #4310, #4365, #4602).
+- Given a hand-written `packages/colonizethis_ai/lib/src/**` Dart file with more than 250 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the run fails naming that file and exits `1` (Refs #4079, #4104, #4239, #4310, #4365, #4602).
+- Given a grandfather allowlist entry for `repo.ai_source_file_size` that names a missing file or a file now at or below 250 physical lines, when the rule runs, then the run fails reporting a stale grandfather entry and exits `1` (Refs #4079, #4104, #4239, #4310, #4365, #4602).
 - Given every `.dart` file under `packages/colonizethis_models/test/**` is at or below 250 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_test_file_size`, then the rule passes without violations and exits `0` (Refs #4068, #4571).
 - Given a `packages/colonizethis_models/test/**` Dart file with more than 250 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_test_file_size`, then the run fails naming that file and exits `1` (Refs #4068, #4571).
 - Given every non-generated `.dart` file under `packages/colonizethis_models/lib/**` is at or below 250 physical lines with an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_lib_physical_file_size`, then the rule passes without violations and exits `0` (Refs #4334, #4571).
