+++ packages/colonizethis_ai/lib/src/planning/diplomacy_planner_run_weight.dart
@@ -1,0 +1,60 @@
+import '../perception/perception_snapshot.dart';
+import 'diplomacy_planner_result.dart';
+import 'expand_phase_planner.dart';
+import 'observer_goal_phase.dart';
+import 'planner_context.dart';
+import 'planning_imports.dart';
+
+int resolveDiplomacyPlannerWeight({
+  required PlannerContext ctx,
+  required AIWorldSnapshot snapshot,
+  required DiplomacyPlannerPass pass,
+}) {
+  var weight = ctx.resolveDiplomacyBaseWeight();
+  if (pass == DiplomacyPlannerPass.declareWarOnly &&
+      snapshot.conquest.provincesToVictory >
+          kConquerScoreFloorProvincesToVictoryThreshold &&
+      weight < 25) {
+    weight = 25;
+  }
+  if (pass == DiplomacyPlannerPass.declareWarOnly &&
+      isStalledOldWorldExpansion(snapshot.conquest.oldWorldProvincesOwned) &&
+      weight < kDiplomacyDeclareWarMinWeightWhenStalled) {
+    weight = kDiplomacyDeclareWarMinWeightWhenStalled;
+  }
+  if (pass == DiplomacyPlannerPass.declareWarOnly &&
+      snapshot.conquest.oldWorldProvincesOwned <=
+          kFewOldWorldProvincesDefendThreshold &&
+      snapshot.conquest.invadableProvinceIdsSorted.isNotEmpty &&
+      weight < kDiplomacyDeclareWarMinWeightWhenStalled + 20) {
+    weight = kDiplomacyDeclareWarMinWeightWhenStalled + 20;
+  }
+  if (pass == DiplomacyPlannerPass.declareWarOnly &&
+      snapshot.conquest.oldWorldProvincesOwned <=
+          kStalledOldWorldProvinceThreshold &&
+      snapshot.conquest.invadableProvinceIdsSorted.isNotEmpty &&
+      weight < kDiplomacyDeclareWarMinWeightWhenStalled + 15) {
+    weight = kDiplomacyDeclareWarMinWeightWhenStalled + 15;
+  }
+  if (pass == DiplomacyPlannerPass.declareWarOnly &&
+      isBelowObserverConquestQuota(snapshot.conquest.oldWorldProvincesOwned) &&
+      snapshot.conquest.invadableProvinceIdsSorted.isNotEmpty &&
+      weight < kDiplomacyDeclareWarMinWeightWhenStalled + 20) {
+    weight = kDiplomacyDeclareWarMinWeightWhenStalled + 20;
+  }
+  if (pass == DiplomacyPlannerPass.declareWarOnly &&
+      hasColonialAcquisitionTargets(snapshot.colonial) &&
+      weight < kDiplomacyDeclareWarMinWeightWhenColonialPressure) {
+    weight = kDiplomacyDeclareWarMinWeightWhenColonialPressure;
+  }
+  if (pass != DiplomacyPlannerPass.declareWarOnly &&
+      (stalledOwExpansionNeedsPeacePass(game: ctx.game, snapshot: snapshot) ||
+          multiFrontNonBlockerGpPeaceTargets(
+            game: ctx.game,
+            snapshot: snapshot,
+          ).isNotEmpty) &&
+      weight < 25) {
+    weight = 25;
+  }
+  return weight;
+}
